split out synonym and props from chado_general as it was getting too large and hard to find the code you want

+++ src/chado_object/feature/chado_feature.py
@@ -912,8 +912,6 @@ class ChadoFeatureObject(ChadoObject):
     def delete_synonym(self, key, bangc=False):
         """Delete synonym.
 
-        Well actually set is_current to false for this entry.
-
         Args:
             key (string): key/field of proforma to get data from.
             bangc (Bool): True if bangc operation.
@@ -926,14 +924,24 @@ class ChadoFeatureObject(ChadoObject):
         else:
             data_list = self.process_data[key]['data']
 
+        cv_name = self.process_data[key]['cv']
+        cvterm_name = self.process_data[key]['cvterm']
+        cvterm = get_cvterm(self.session, cv_name, cvterm_name)
+        if not cvterm:
+            message = "Unable to find cvterm {} for Cv {}.".format(cvterm_name, cv_name)
+            self.critical_error(self.process_data[key]['data'], message)
+            return None
+
         if bangc:
             self.session.query(FeatureSynonym).\
                 filter(FeatureSynonym.pub_id == self.pub.pub_id,
+                       FeatureSynonym.type_id == cvterm.cvterm_id,
                        FeatureSynonym.feature_id == self.feature.feature_id).delete()
         else:
             for data in data_list:
                 synonyms = self.session.query(Synonym).\
-                    filter(Synonym.name == data[FIELD_VALUE])
+                    filter(Synonym.name == sgml_to_plain_text(data[FIELD_VALUE]),
+                           FeatureSynonym.type_id == cvterm.cvterm_id)
                 syn_count = 0
                 f_syn_count = 0
                 for syn in synonyms:
